+++ libft/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: bpoyet <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/11 12:32:37 by bpoyet            #+#    #+#             */
-/*   Updated: 2023/08/11 12:51:49 by bpoyet           ###   ########.fr       */
+/*   Updated: 2023/11/08 18:00:03 by bpoyet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *str)
+#include "libft.h"
+
+size_t ft_strlen(char *str)
 {
-	int	i;
+	int i;
 
 	i = 0;
 	while (str[i] != '\0')
